refactor: add flush_test.rs
