+++ src/mito2/src/engine/tests.rs
@@ -12,25 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Tests for mito engine.
+//! Basic tests for mito engine.
 
 use std::collections::HashMap;
 
-use api::helper::ColumnDataTypeWrapper;
-use api::v1::value::ValueData;
-use api::v1::{ColumnSchema, Row, Rows, SemanticType};
+use api::v1::Rows;
 use common_recordbatch::RecordBatches;
-use store_api::metadata::ColumnMetadata;
-use store_api::region_request::{
-    RegionCreateRequest, RegionDeleteRequest, RegionFlushRequest, RegionOpenRequest,
-    RegionPutRequest,
-};
+use store_api::region_request::RegionOpenRequest;
 use store_api::storage::RegionId;
 
 use super::*;
 use crate::error::Error;
 use crate::region::version::VersionControlData;
-use crate::test_util::{CreateRequestBuilder, FlushListener, MockWriteBufferManager, TestEnv};
+use crate::test_util::{
+    build_delete_rows_for_key, build_rows, build_rows_for_key, delete_rows, delete_rows_schema,
+    put_rows, rows_schema, CreateRequestBuilder, TestEnv,
+};
 
 #[tokio::test]
 async fn test_engine_new_stop() {
@@ -57,54 +54,6 @@ async fn test_engine_new_stop() {
         matches!(err, Error::WorkerStopped { .. }),
         "unexpected err: {err}"
     );
-}
-
-fn column_metadata_to_column_schema(metadata: &ColumnMetadata) -> api::v1::ColumnSchema {
-    api::v1::ColumnSchema {
-        column_name: metadata.column_schema.name.clone(),
-        datatype: ColumnDataTypeWrapper::try_from(metadata.column_schema.data_type.clone())
-            .unwrap()
-            .datatype() as i32,
-        semantic_type: metadata.semantic_type as i32,
-    }
-}
-
-fn build_rows(start: usize, end: usize) -> Vec<Row> {
-    (start..end)
-        .map(|i| api::v1::Row {
-            values: vec![
-                api::v1::Value {
-                    value_data: Some(ValueData::StringValue(i.to_string())),
-                },
-                api::v1::Value {
-                    value_data: Some(ValueData::F64Value(i as f64)),
-                },
-                api::v1::Value {
-                    value_data: Some(ValueData::TsMillisecondValue(i as i64 * 1000)),
-                },
-            ],
-        })
-        .collect()
-}
-
-fn rows_schema(request: &RegionCreateRequest) -> Vec<ColumnSchema> {
-    request
-        .column_metadatas
-        .iter()
-        .map(column_metadata_to_column_schema)
-        .collect::<Vec<_>>()
-}
-
-async fn put_rows(engine: &MitoEngine, region_id: RegionId, rows: Rows) {
-    let num_rows = rows.rows.len();
-    let output = engine
-        .handle_request(region_id, RegionRequest::Put(RegionPutRequest { rows }))
-        .await
-        .unwrap();
-    let Output::AffectedRows(rows_inserted) = output else {
-        unreachable!()
-    };
-    assert_eq!(num_rows, rows_inserted);
 }
 
 #[tokio::test]
@@ -193,8 +142,6 @@ async fn test_region_replay() {
     engine.stop().await.unwrap();
 }
 
-// TODO(yingwen): build_rows() only generate one point for each series. We need to add tests
-// for series with multiple points and other cases.
 #[tokio::test]
 async fn test_write_query_region() {
     let mut env = TestEnv::new();
@@ -228,66 +175,6 @@ async fn test_write_query_region() {
 | 2     | 2.0     | 1970-01-01T00:00:02 |
 +-------+---------+---------------------+";
     assert_eq!(expected, batches.pretty_print().unwrap());
-}
-
-/// Build rows to put for specific `key`.
-fn build_rows_for_key(key: &str, start: usize, end: usize, value_start: usize) -> Vec<Row> {
-    (start..end)
-        .enumerate()
-        .map(|(idx, ts)| api::v1::Row {
-            values: vec![
-                api::v1::Value {
-                    value_data: Some(ValueData::StringValue(key.to_string())),
-                },
-                api::v1::Value {
-                    value_data: Some(ValueData::F64Value((value_start + idx) as f64)),
-                },
-                api::v1::Value {
-                    value_data: Some(ValueData::TsMillisecondValue(ts as i64 * 1000)),
-                },
-            ],
-        })
-        .collect()
-}
-
-/// Build rows to delete for specific `key`.
-fn build_delete_rows_for_key(key: &str, start: usize, end: usize) -> Vec<Row> {
-    (start..end)
-        .map(|ts| api::v1::Row {
-            values: vec![
-                api::v1::Value {
-                    value_data: Some(ValueData::StringValue(key.to_string())),
-                },
-                api::v1::Value {
-                    value_data: Some(ValueData::TsMillisecondValue(ts as i64 * 1000)),
-                },
-            ],
-        })
-        .collect()
-}
-
-fn delete_rows_schema(request: &RegionCreateRequest) -> Vec<ColumnSchema> {
-    request
-        .column_metadatas
-        .iter()
-        .filter(|col| col.semantic_type != SemanticType::Field)
-        .map(column_metadata_to_column_schema)
-        .collect::<Vec<_>>()
-}
-
-async fn delete_rows(engine: &MitoEngine, region_id: RegionId, rows: Rows) {
-    let num_rows = rows.rows.len();
-    let output = engine
-        .handle_request(
-            region_id,
-            RegionRequest::Delete(RegionDeleteRequest { rows }),
-        )
-        .await
-        .unwrap();
-    let Output::AffectedRows(rows_inserted) = output else {
-        unreachable!()
-    };
-    assert_eq!(num_rows, rows_inserted);
 }
 
 #[tokio::test]
@@ -400,109 +287,6 @@ async fn test_put_overwrite() {
 | b     | 3.0     | 1970-01-01T00:00:00 |
 | b     | 1.0     | 1970-01-01T00:00:01 |
 | b     | 4.0     | 1970-01-01T00:00:02 |
-+-------+---------+---------------------+";
-    assert_eq!(expected, batches.pretty_print().unwrap());
-}
-
-#[tokio::test]
-async fn test_manual_flush() {
-    let mut env = TestEnv::new();
-    let engine = env.create_engine(MitoConfig::default()).await;
-
-    let region_id = RegionId::new(1, 1);
-    let request = CreateRequestBuilder::new().build();
-
-    let column_schemas = rows_schema(&request);
-    engine
-        .handle_request(region_id, RegionRequest::Create(request))
-        .await
-        .unwrap();
-
-    let rows = Rows {
-        schema: column_schemas,
-        rows: build_rows(0, 3),
-    };
-    put_rows(&engine, region_id, rows).await;
-
-    let Output::AffectedRows(rows) = engine
-        .handle_request(region_id, RegionRequest::Flush(RegionFlushRequest {}))
-        .await
-        .unwrap()
-    else {
-        unreachable!()
-    };
-    assert_eq!(0, rows);
-
-    let request = ScanRequest::default();
-    let scanner = engine.handle_query(region_id, request).unwrap();
-    assert_eq!(1, scanner.num_files());
-    let stream = scanner.scan().await.unwrap();
-    let batches = RecordBatches::try_collect(stream).await.unwrap();
-    let expected = "\
-+-------+---------+---------------------+
-| tag_0 | field_0 | ts                  |
-+-------+---------+---------------------+
-| 0     | 0.0     | 1970-01-01T00:00:00 |
-| 1     | 1.0     | 1970-01-01T00:00:01 |
-| 2     | 2.0     | 1970-01-01T00:00:02 |
-+-------+---------+---------------------+";
-    assert_eq!(expected, batches.pretty_print().unwrap());
-}
-
-#[tokio::test]
-async fn test_flush_engine() {
-    let mut env = TestEnv::new();
-    let write_buffer_manager = Arc::new(MockWriteBufferManager::default());
-    let listener = Arc::new(FlushListener::new());
-    let engine = env
-        .create_engine_with(
-            MitoConfig::default(),
-            write_buffer_manager.clone(),
-            Some(listener.clone()),
-        )
-        .await;
-
-    let region_id = RegionId::new(1, 1);
-    let request = CreateRequestBuilder::new().build();
-
-    let column_schemas = rows_schema(&request);
-    engine
-        .handle_request(region_id, RegionRequest::Create(request))
-        .await
-        .unwrap();
-
-    // Prepares rows for flush.
-    let rows = Rows {
-        schema: column_schemas.clone(),
-        rows: build_rows_for_key("a", 0, 2, 0),
-    };
-    put_rows(&engine, region_id, rows).await;
-
-    write_buffer_manager.set_should_flush(true);
-
-    // Writes and triggers flush.
-    let rows = Rows {
-        schema: column_schemas.clone(),
-        rows: build_rows_for_key("b", 0, 2, 0),
-    };
-    put_rows(&engine, region_id, rows).await;
-
-    // Wait until flush is finished.
-    listener.wait().await;
-
-    let request = ScanRequest::default();
-    let scanner = engine.handle_query(region_id, request).unwrap();
-    assert_eq!(1, scanner.num_files());
-    let stream = scanner.scan().await.unwrap();
-    let batches = RecordBatches::try_collect(stream).await.unwrap();
-    let expected = "\
-+-------+---------+---------------------+
-| tag_0 | field_0 | ts                  |
-+-------+---------+---------------------+
-| a     | 0.0     | 1970-01-01T00:00:00 |
-| a     | 1.0     | 1970-01-01T00:00:01 |
-| b     | 0.0     | 1970-01-01T00:00:00 |
-| b     | 1.0     | 1970-01-01T00:00:01 |
 +-------+---------+---------------------+";
     assert_eq!(expected, batches.pretty_print().unwrap());
 }
